requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
pydantic-settings
python-jose
passlib
python-dotenv
